backend/requirements.txt
inquirer
firebase-admin 
google-auth 
google-auth-oauthlib 
google-auth-httplib2
flask
flask_cors
flask_session
python-dotenv
gunicorn